+++ src/template/js/src/client/api/beer.js
@@ -53,3 +53,16 @@ export const addBeer = async ({ abv, brewer, description, name, type }) => {
     console.error(error);
   }
 };
+
+export const deleteBeer = async (id) => {
+  try {
+    await fetch(`/beer/${id}`, {
+      headers: {
+        "Content-Type": "application/json",
+      },
+      method: "DELETE",
+    });
+  } catch (error) {
+    console.error(error);
+  }
+};