+++ src/template/js/src/client/components/BeerDisplayer.js
@@ -1,6 +1,12 @@
 import * as React from "react";
 
-const BeerDisplayer = ({ beers, setIsModalOpen, setModalContent }) => {
+const BeerDisplayer = ({
+  beers,
+  refreshBeers,
+  setCurrentBeer,
+  setIsModalOpen,
+  setModalContent,
+}) => {
   return (
     <div>
       <div id="beer-table-heading">
@@ -35,7 +41,15 @@ const BeerDisplayer = ({ beers, setIsModalOpen, setModalContent }) => {
               <td>{beer.type}</td>
               <td className="column6 action-column">
                 <button>Edit</button>
-                <button>Delete</button>
+                <button
+                  onClick={() => {
+                    setModalContent("deleteBeerForm");
+                    setCurrentBeer(beer);
+                    setIsModalOpen(true);
+                  }}
+                >
+                  Delete
+                </button>
               </td>
             </tr>
           ))}