+++ src/template/ts/src/client/components/Modal.tsx
@@ -1,18 +1,24 @@
 import * as React from "react";
 import NewBeerForm from "./forms/NewBeerForm";
+import DeleteBeerForm from "./forms/DeleteBeerForm";
+import { Beer } from "../types";
 
 export interface ModalProps {
+  currentBeer?: Beer;
   isModalOpen: boolean;
   modalContent?: any;
   refreshBeers: Function;
+  setCurrentBeer: Function;
   setIsModalOpen: Function;
   setModalContent: Function;
 }
 
 const Modal: React.SFC<ModalProps> = ({
+  currentBeer,
   isModalOpen,
   modalContent,
   refreshBeers,
+  setCurrentBeer,
   setIsModalOpen,
   setModalContent,
 }) => {
@@ -37,10 +43,21 @@ const Modal: React.SFC<ModalProps> = ({
             x
           </button>
         </div>
-        <NewBeerForm
-          refreshBeers={refreshBeers}
-          setIsModalOpen={setIsModalOpen}
-        />
+        {modalContent === "newBeerForm" && (
+          <NewBeerForm
+            refreshBeers={refreshBeers}
+            setIsModalOpen={setIsModalOpen}
+          />
+        )}
+        {modalContent === "deleteBeerForm" && (
+          <DeleteBeerForm
+            currentBeer={currentBeer}
+            refreshBeers={refreshBeers}
+            setCurrentBeer={setCurrentBeer}
+            setIsModalOpen={setIsModalOpen}
+            setModalContent={setModalContent}
+          />
+        )}
       </div>
     </>
   );